backend/requirements.txt
# Existing dependencies
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
black==23.11.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Faker==19.13.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
isort==5.12.0
motor==3.3.2
mypy_extensions==1.1.0
openpyxl==3.1.2
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pymongo==4.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.2
reportlab==4.0.7
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tomli==2.2.1
typing_extensions==4.14.1
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1

# AI & NLP Dependencies untuk Luna Chatbot
torch>=1.9.0,<2.1.0
transformers>=4.20.0,<5.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
numpy>=1.21.0,<1.25.0
regex>=2022.0.0
tokenizers>=0.13.0
huggingface-hub>=0.14.0
safetensors>=0.3.0